refactor(browser): share context-menu suppression and response handling
