+++ composeApp/src/desktopTest/kotlin/ai/rever/boss/plugin/browser/PopupWindowContextMenuTest.kt
@@ -432,5 +432,4 @@ class PopupWindowContextMenuTest {
         callback.on(params, tell)
         assertTrue(tell.isClosed)
     }
-
 }
